+++ src/journals/settings/ui/sections/NoteCreationSection.vue
@@ -74,6 +74,9 @@ function applyDateFormatRecommendation(): void {
           <template v-if="invertibility.kind === 'non-invertible'">
             {{ m.journal_edit_name_template_invertibility_warning(invertibility) }}
           </template>
+          <template v-else-if="invertibility.kind === 'coarse-date'">
+            {{ m.journal_edit_name_template_coarse_date_warning() }}
+          </template>
           <template v-else-if="invertibility.kind === 'cyclic-top'">
             {{ m.journal_edit_name_template_cyclic_top_warning() }}
           </template>